requirements.txt
# Data source adapters (install only what you need)
kbcstorage>=0.9.0          # For Keboola adapter

# Data processing
# pandas - hlavní knihovna pro práci s tabulkovými daty
# pyarrow - podpora pro Parquet formát a rychlé operace
# pytz - timezone support required by DuckDB for reading timezone-aware Parquet columns
pandas>=2.0.0
pyarrow>=12.0.0
pytz>=2024.1

# Analytická databáze
# DuckDB - in-process SQL OLAP databáze pro analytické dotazy
duckdb>=0.9.0

# Konfigurace
# python-dotenv - načítání environment variables z .env souborů
# pyyaml - parsování YAML konfigurace z data_description.md
python-dotenv>=1.0.0
pyyaml>=6.0

# Progress tracking a logging
# tqdm - progress bary pro dlouhotrvající operace (download, sync)
tqdm>=4.65.0

# Web application (Google SSO portal)
# flask - web framework pro self-service portal
# authlib - OAuth 2.0 / OpenID Connect knihovna pro Google SSO
# gunicorn - WSGI server pro production deployment
flask>=3.0.0
authlib>=1.3.0
gunicorn>=21.0.0

# Telegram notification bot
# httpx - async HTTP client for Telegram API and unix socket communication
# aiohttp - async HTTP server for bot's internal send API
httpx>=0.27.0
aiohttp>=3.9.0
matplotlib>=3.8.0
numpy>=1.24.0

# Desktop app authentication
# PyJWT - JWT token creation and validation for desktop app auth
PyJWT>=2.8.0

# Password authentication for external users
# argon2-cffi - modern password hashing algorithm (Argon2id)
# sendgrid - email service for setup/reset links
argon2-cffi>=23.1.0
sendgrid>=6.11.0

# Corporate Memory knowledge extraction
# anthropic - Claude API client for HAIKU-based knowledge extraction
anthropic>=0.39.0
